commit with comments in backend
