+++ backend/server.js
@@ -7,6 +7,7 @@ import connectdb from "./db/connectDb.js";
 import cookieParser from "cookie-parser";
 import auth from "./routes/auth.js";
 import post from "./routes/post.js";
+import likeComment from "./routes/likeandcomment.js";
 // import Post from "./models/post.js";
 import Post from "./models/post.js";
 import Comment from "./models/comment.js";
@@ -59,197 +60,198 @@ app.options("*", cors());
 // Routes
 app.use("/api", auth);
 app.use("/api", post);
+app.use("/api", likeComment);
 
 //****************************************** */
 
-// Backend route for handling reactions
-app.post('/api/posts/:postId/react', auth, async (req, res) => {
-  const { postId } = req.params;
-  const { reactionType } = req.body;
-  const userId = req.user.id; // Assuming auth middleware adds `user` to `req`
+// // Backend route for handling reactions
+// app.post('/api/posts/:postId/react', auth, async (req, res) => {
+//   const { postId } = req.params;
+//   const { reactionType } = req.body;
+//   const userId = req.user.id; // Assuming auth middleware adds `user` to `req`
 
-  try {
-    const post = await Post.findById(postId);
-    if (!post) {
-      return res.status(404).json({ message: "Post not found" });
-    }
+//   try {
+//     const post = await Post.findById(postId);
+//     if (!post) {
+//       return res.status(404).json({ message: "Post not found" });
+//     }
 
-    // Check if the user already has a reaction
-    const existingReactionIndex = post.reactions.findIndex(
-      (r) => r.user.toString() === userId
-    );
+//     // Check if the user already has a reaction
+//     const existingReactionIndex = post.reactions.findIndex(
+//       (r) => r.user.toString() === userId
+//     );
 
-    if (existingReactionIndex !== -1) {
-      // If the user already has a reaction, update it
-      post.reactions[existingReactionIndex].type = reactionType;
-    } else {
-      // If the user doesn't have a reaction, add a new one
-      post.reactions.push({ user: userId, type: reactionType });
-    }
+//     if (existingReactionIndex !== -1) {
+//       // If the user already has a reaction, update it
+//       post.reactions[existingReactionIndex].type = reactionType;
+//     } else {
+//       // If the user doesn't have a reaction, add a new one
+//       post.reactions.push({ user: userId, type: reactionType });
+//     }
 
-    await post.save();
+//     await post.save();
 
-    // Count reactions by type
-    const reactionCounts = post.reactions.reduce((acc, reaction) => {
-      acc[reaction.type] = (acc[reaction.type] || 0) + 1;
-      return acc;
-    }, {});
+//     // Count reactions by type
+//     const reactionCounts = post.reactions.reduce((acc, reaction) => {
+//       acc[reaction.type] = (acc[reaction.type] || 0) + 1;
+//       return acc;
+//     }, {});
 
-    res.json({ message: "Reaction updated", reactionCounts });
-  } catch (err) {
-    console.error("Error reacting to post", err);
-    res.status(500).json({ message: "Internal server error" });
-  }
-});
+//     res.json({ message: "Reaction updated", reactionCounts });
+//   } catch (err) {
+//     console.error("Error reacting to post", err);
+//     res.status(500).json({ message: "Internal server error" });
+//   }
+// });
 
-app.get('/api/posts/:postId/reactions', auth, async (req, res) => {
-  const { postId } = req.params;
+// app.get('/api/posts/:postId/reactions', auth, async (req, res) => {
+//   const { postId } = req.params;
 
-  try {
-    const post = await Post.findById(postId);
-    if (!post) {
-      return res.status(404).json({ message: "Post not found" });
-    }
+//   try {
+//     const post = await Post.findById(postId);
+//     if (!post) {
+//       return res.status(404).json({ message: "Post not found" });
+//     }
 
-    // Count reactions by type
-    const reactionCounts = post.reactions.reduce((acc, reaction) => {
-      acc[reaction.type] = (acc[reaction.type] || 0) + 1;
-      return acc;
-    }, {});
+//     // Count reactions by type
+//     const reactionCounts = post.reactions.reduce((acc, reaction) => {
+//       acc[reaction.type] = (acc[reaction.type] || 0) + 1;
+//       return acc;
+//     }, {});
 
-    // Find the user's reaction
-    const userReaction = post.reactions.find(
-      (r) => r.user.toString() === req.user.id
-    )?.type;
+//     // Find the user's reaction
+//     const userReaction = post.reactions.find(
+//       (r) => r.user.toString() === req.user.id
+//     )?.type;
 
-    res.json({ reactionCounts, userReaction });
-  } catch (err) {
-    console.error("Error fetching reactions", err);
-    res.status(500).json({ message: "Internal server error" });
-  }
-});
-
-
+//     res.json({ reactionCounts, userReaction });
+//   } catch (err) {
+//     console.error("Error fetching reactions", err);
+//     res.status(500).json({ message: "Internal server error" });
+//   }
+// });
 
 
-// Add a new comment
-app.post('/api/posts/:postId/comments', async (req, res) => {
-  try {
-    const { user, content, parentComment } = req.body;
-    const { postId } = req.params;
 
-    console.log("Received payload:", { user, content, postId, parentComment }); // Debugging: Log the received payload
 
-    // Validate user ID
-    if (!mongoose.Types.ObjectId.isValid(user)) {
-      console.error("Invalid user ID:", user); // Debugging: Log invalid user ID
-      return res.status(400).json({ error: 'Invalid user ID' });
-    }
+// // Add a new comment
+// app.post('/api/posts/:postId/comments', async (req, res) => {
+//   try {
+//     const { user, content, parentComment } = req.body;
+//     const { postId } = req.params;
 
-    // Validate post ID
-    if (!mongoose.Types.ObjectId.isValid(postId)) {
-      console.error("Invalid post ID:", postId); // Debugging: Log invalid post ID
-      return res.status(400).json({ error: 'Invalid post ID' });
-    }
+//     console.log("Received payload:", { user, content, postId, parentComment }); // Debugging: Log the received payload
 
-    // Ensure content is not empty
-    if (!content || content.trim() === '') {
-      console.error("Empty content:", content); // Debugging: Log empty content
-      return res.status(400).json({ error: 'Comment content is required' });
-    }
+//     // Validate user ID
+//     if (!mongoose.Types.ObjectId.isValid(user)) {
+//       console.error("Invalid user ID:", user); // Debugging: Log invalid user ID
+//       return res.status(400).json({ error: 'Invalid user ID' });
+//     }
 
-    const comment = new Comment({
-      user: new mongoose.Types.ObjectId(user),
-      post: new mongoose.Types.ObjectId(postId),
-      content,
-      parentComment: parentComment ? new mongoose.Types.ObjectId(parentComment) : null,
-    });
+//     // Validate post ID
+//     if (!mongoose.Types.ObjectId.isValid(postId)) {
+//       console.error("Invalid post ID:", postId); // Debugging: Log invalid post ID
+//       return res.status(400).json({ error: 'Invalid post ID' });
+//     }
 
-    await comment.save();
-    res.status(201).json(comment);
-  } catch (error) {
-    console.error('Error adding comment:', error);
-    res.status(500).json({ error: 'Internal Server Error' });
-  }
-});
+//     // Ensure content is not empty
+//     if (!content || content.trim() === '') {
+//       console.error("Empty content:", content); // Debugging: Log empty content
+//       return res.status(400).json({ error: 'Comment content is required' });
+//     }
 
-// Like a comment
-app.post('/api/comments/:commentId/like', auth, async (req, res) => {
-  try {
-    const { commentId } = req.params;
-    const comment = await Comment.findById(commentId);
-    if (!comment) return res.status(404).json({ error: "Comment not found" });
+//     const comment = new Comment({
+//       user: new mongoose.Types.ObjectId(user),
+//       post: new mongoose.Types.ObjectId(postId),
+//       content,
+//       parentComment: parentComment ? new mongoose.Types.ObjectId(parentComment) : null,
+//     });
 
-    comment.likes = (comment.likes || 0) + 1;
-    await comment.save();
-    res.json(comment);
-  } catch (error) {
-    console.error("Error liking comment:", error);
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-});
+//     await comment.save();
+//     res.status(201).json(comment);
+//   } catch (error) {
+//     console.error('Error adding comment:', error);
+//     res.status(500).json({ error: 'Internal Server Error' });
+//   }
+// });
 
-// Edit a comment
-app.put('/api/comments/:commentId', auth, async (req, res) => {
-  try {
-    const { commentId } = req.params;
-    const { content } = req.body;
-    const comment = await Comment.findByIdAndUpdate(
-      commentId,
-      { content },
-      { new: true }
-    );
-    res.json(comment);
-  } catch (error) {
-    console.error("Error editing comment:", error);
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-});
+// // Like a comment
+// app.post('/api/comments/:commentId/like', auth, async (req, res) => {
+//   try {
+//     const { commentId } = req.params;
+//     const comment = await Comment.findById(commentId);
+//     if (!comment) return res.status(404).json({ error: "Comment not found" });
 
-// Delete a comment
-// Your routes
-app.delete("/api/comments/:commentId", async (req, res) => {
-  try {
-    const { commentId } = req.params;
-    await Comment.findByIdAndDelete(commentId);
-    res.status(204).send();
-  } catch (error) {
-    console.error("Error deleting comment:", error);
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-});
+//     comment.likes = (comment.likes || 0) + 1;
+//     await comment.save();
+//     res.json(comment);
+//   } catch (error) {
+//     console.error("Error liking comment:", error);
+//     res.status(500).json({ error: "Internal Server Error" });
+//   }
+// });
 
-// Get all comments for a post
-app.get('/api/posts/:postId/comments', auth, async (req, res) => {
-  try {
-    const { postId } = req.params;
+// // Edit a comment
+// app.put('/api/comments/:commentId', auth, async (req, res) => {
+//   try {
+//     const { commentId } = req.params;
+//     const { content } = req.body;
+//     const comment = await Comment.findByIdAndUpdate(
+//       commentId,
+//       { content },
+//       { new: true }
+//     );
+//     res.json(comment);
+//   } catch (error) {
+//     console.error("Error editing comment:", error);
+//     res.status(500).json({ error: "Internal Server Error" });
+//   }
+// });
 
-    // Validate post ID
-    if (!mongoose.Types.ObjectId.isValid(postId)) {
-      return res.status(400).json({ error: 'Invalid post ID' });
-    }
+// // Delete a comment
+// // Your routes
+// app.delete("/api/comments/:commentId", async (req, res) => {
+//   try {
+//     const { commentId } = req.params;
+//     await Comment.findByIdAndDelete(commentId);
+//     res.status(204).send();
+//   } catch (error) {
+//     console.error("Error deleting comment:", error);
+//     res.status(500).json({ error: "Internal Server Error" });
+//   }
+// });
 
-    const comments = await Comment.find({ post: postId })
-      .populate('user', 'name email profilePicture'); // Only fetch necessary fields
+// // Get all comments for a post
+// app.get('/api/posts/:postId/comments', auth, async (req, res) => {
+//   try {
+//     const { postId } = req.params;
 
-    res.json(comments);
-  } catch (error) {
-    console.error('Error fetching comments:', error);
-    res.status(500).json({ error: 'Internal Server Error' });
-  }
-});
+//     // Validate post ID
+//     if (!mongoose.Types.ObjectId.isValid(postId)) {
+//       return res.status(400).json({ error: 'Invalid post ID' });
+//     }
 
-// Backend route to get parent comment count
-app.get("/api/posts/:postId/parent-comments-count", async (req, res) => {
-  try {
-    const postId = req.params.postId;
-    const count = await Comment.countDocuments({ post: postId, parentComment: null });
-    res.json({ count });
-  } catch (error) {
-    console.error("Error fetching parent comment count:", error);
-    res.status(500).json({ message: "Internal server error" });
-  }
-});
+//     const comments = await Comment.find({ post: postId })
+//       .populate('user', 'name email profilePicture'); // Only fetch necessary fields
+
+//     res.json(comments);
+//   } catch (error) {
+//     console.error('Error fetching comments:', error);
+//     res.status(500).json({ error: 'Internal Server Error' });
+//   }
+// });
+
+// // Backend route to get parent comment count
+// app.get("/api/posts/:postId/parent-comments-count", async (req, res) => {
+//   try {
+//     const postId = req.params.postId;
+//     const count = await Comment.countDocuments({ post: postId, parentComment: null });
+//     res.json({ count });
+//   } catch (error) {
+//     console.error("Error fetching parent comment count:", error);
+//     res.status(500).json({ message: "Internal server error" });
+//   }
+// });
 
 
 //********************************************************************** */
